Move TabBarCoordinatorParameters.
